+++ packages/tsurlfilter/CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `$to` modifier in the MV3 converter.
 - Support for `$method` modifier in the MV3 converter.
 
+### Fixed
+- Bad conversion of `allowAllRequests` rules.
+
 
 ## [2.1.6] - 2023-08-04